AutoFillGroup: initial commit
