+++ packages/flet/lib/src/controls/autofill_group.dart
@@ -1,37 +1,41 @@
+import 'package:flet/src/utils/autofill.dart';
 import 'package:flutter/material.dart';
 
 import '../models/control.dart';
 import 'create_control.dart';
 import 'error.dart';
 
-class RadioGroupControl extends StatelessWidget {
+class AutofillGroupControl extends StatelessWidget {
   final Control? parent;
   final Control control;
   final List<Control> children;
   final bool parentDisabled;
   final bool? parentAdaptive;
 
-  const RadioGroupControl(
+  const AutofillGroupControl(
       {super.key,
-      this.parent,
+      required this.parent,
       required this.control,
       required this.children,
       required this.parentDisabled,
-      required this.parentAdaptive});
+      this.parentAdaptive});
 
   @override
   Widget build(BuildContext context) {
-    debugPrint("RadioGroupControl build: ${control.id}");
+    debugPrint("AutofillGroup build: ${control.id}");
 
     var contentCtrls =
         children.where((c) => c.name == "content" && c.isVisible);
     bool disabled = control.isDisabled || parentDisabled;
 
     if (contentCtrls.isEmpty) {
-      return const ErrorControl("RadioGroup control has no content.");
+      return const ErrorControl("AutofillGroup control has no content.");
     }
 
-    return createControl(control, contentCtrls.first.id, disabled,
-        parentAdaptive: parentAdaptive);
+    return AutofillGroup(
+        onDisposeAction: parseAutofillContextAction(
+            control.attrString("disposeAction"), AutofillContextAction.commit)!,
+        child: createControl(control, contentCtrls.first.id, disabled,
+            parentAdaptive: parentAdaptive));
   }
 }
