+++ sdk/python/packages/flet-core/src/flet_core/radio_group.py
@@ -66,10 +66,10 @@ class RadioGroup(Control):
         return "radiogroup"
 
     def _get_children(self):
-        if self.__content is None:
-            return []
-        self.__content._set_attr_internal("n", "content")
-        return [self.__content]
+        if isinstance(self.__content, Control):
+            self.__content._set_attr_internal("n", "content")
+            return [self.__content]
+        return []
 
     # value
     @property
